requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
black==23.11.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
distlib==0.3.7
exceptiongroup==1.1.3
fastapi==0.104.1
FastAPI-SQLAlchemy==0.2.1
filelock==3.13.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pipenv==2023.11.15
platformdirs==3.11.0
psycopg2==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
regex==2023.10.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
virtualenv==20.24.6
virtualenv-clone==0.5.7
